+++ sdks/java/io/kinesis/src/main/java/org/apache/beam/sdk/io/kinesis/GetKinesisRecordsResult.java
@@ -1,0 +1,54 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.beam.sdk.io.kinesis;
+
+import static com.google.common.collect.Lists.transform;
+import com.google.common.base.Function;
+
+import com.amazonaws.services.kinesis.clientlibrary.types.UserRecord;
+import java.util.List;
+import javax.annotation.Nullable;
+
+/***
+ * Represents the output of 'get' operation on Kinesis stream.
+ */
+class GetKinesisRecordsResult {
+    private final List<KinesisRecord> records;
+    private final String nextShardIterator;
+
+    public GetKinesisRecordsResult(List<UserRecord> records, String nextShardIterator,
+                                   final String streamName, final String shardId) {
+        this.records = transform(records, new Function<UserRecord, KinesisRecord>() {
+            @Nullable
+            @Override
+            public KinesisRecord apply(@Nullable UserRecord input) {
+                assert input != null;  // to make FindBugs happy
+                return new KinesisRecord(input, streamName, shardId);
+            }
+        });
+        this.nextShardIterator = nextShardIterator;
+    }
+
+    public List<KinesisRecord> getRecords() {
+        return records;
+    }
+
+    public String getNextShardIterator() {
+        return nextShardIterator;
+    }
+}
