+++ sdks/java/io/kinesis/src/main/java/org/apache/beam/sdk/io/kinesis/KinesisIO.java
@@ -18,11 +18,9 @@
 package org.apache.beam.sdk.io.kinesis;
 
 
-import org.apache.beam.sdk.io.kinesis.client.KinesisClientProvider;
-import org.apache.beam.sdk.io.kinesis.client.response.KinesisRecord;
-import org.apache.beam.sdk.io.kinesis.source.KinesisSource;
-import org.apache.beam.sdk.io.kinesis.source.checkpoint.StartingPoint;
 import org.apache.beam.sdk.transforms.PTransform;
+
+import static com.google.common.base.Preconditions.checkNotNull;
 
 import com.amazonaws.auth.AWSCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
@@ -31,22 +29,94 @@ import com.amazonaws.regions.Regions;
 import com.amazonaws.services.kinesis.AmazonKinesis;
 import com.amazonaws.services.kinesis.AmazonKinesisClient;
 import com.amazonaws.services.kinesis.clientlibrary.lib.worker.InitialPositionInStream;
-import java.util.Date;
+import org.joda.time.Instant;
 
 /**
  * {@link PTransform}s for reading from
  * <a href="https://aws.amazon.com/kinesis/">Kinesis</a> streams.
+ *
+ * <h3>Usage</h3>
+ *
+ * <p>Main class you're going to operate is called {@link KinesisIO}.
+ * It follows the usage conventions laid out by other *IO classes like
+ * BigQueryIO or PubsubIOLet's see how you can set up a simple Pipeline, which reads from Kinesis:
+ *
+ * <pre>{@code}
+ * p.
+ *   apply(KinesisIO.Read.
+ *     from("streamName", InitialPositionInStream.LATEST).
+ *     using("AWS_KEY", _"AWS_SECRET", STREAM_REGION).
+ *     apply( ... ) // other transformations
+ *</pre>
+ * </p>
+ *
+ * <p>
+ * As you can see you need to provide 3 things:
+ * <ul>
+ *   <li>name of the stream you're going to read</li>
+ *   <li>position in the stream where reading should start. There are two options:</li>
+ *   <ul>
+ *     <li>{@link InitialPositionInStream#LATEST} - reading will begin from end of the stream</li>
+ *     <li>{@link InitialPositionInStream#TRIM_HORIZON} - reading will begin at
+ *        the very beginning of the stream</li>
+ *   </ul>
+ *   <li>data used to initialize {@link AmazonKinesis} client></li>
+ *   <ul>
+ *     <li>credentials (aws key, aws secret)</li>
+ *    <li>region where the stream is located</li>
+ *   </ul>
+ * </ul>
+ * </p>
+ *
+ * <p>In case when you want to set up {@link AmazonKinesis} client by your own
+ * (for example if you're using more sophisticated authorization methods like Amazon STS, etc.)
+ * you can do it by implementing {@link KinesisClientProvider} class:
+ *
+ * <pre>{@code}
+ * public class MyCustomKinesisClientProvider implements KinesisClientProvider {
+ *   @Override
+ *   public AmazonKinesis get() {
+ *     // set up your client here
+ *   }
+ * }
+ * </pre>
+ *
+ * Usage is pretty straightforward:
+ *
+ * <pre>{@code}
+ * p.
+ *   apply(KinesisIO.Read.
+ *    from("streamName", InitialPositionInStream.LATEST).
+ *    using(MyCustomKinesisClientProvider()).
+ *    apply( ... ) // other transformations
+ * </pre>
+ * </p>
+ *
+ * <p>There’s also possibility to start reading using arbitrary point in time -
+ * in this case you need to provide {@link Instant} object:
+ *
+ * <pre>{@code}
+ * p.
+ *   apply(KinesisIO.Read.
+ *     from("streamName", instant).
+ *     using(MyCustomKinesisClientProvider()).
+ *     apply( ... ) // other transformations
+ * </pre>
+ * </p>
+ *
  */
-public class KinesisIO {
+public final class KinesisIO {
     /***
      * A {@link PTransform} that reads from a Kinesis stream.
      */
-    public static class Read {
+    public static final class Read {
 
         private final String streamName;
         private final StartingPoint initialPosition;
 
         private Read(String streamName, StartingPoint initialPosition) {
+            checkNotNull(streamName);
+            checkNotNull(initialPosition);
             this.streamName = streamName;
             this.initialPosition = initialPosition;
         }
@@ -58,7 +128,7 @@ public class KinesisIO {
             return new Read(streamName, new StartingPoint(initialPosition));
         }
 
-        public static Read from(String streamName, Date initialTimestamp) {
+        public static Read from(String streamName, Instant initialTimestamp) {
             return new Read(streamName, new StartingPoint(initialTimestamp));
         }
 
@@ -87,7 +157,7 @@ public class KinesisIO {
             return using(new BasicKinesisProvider(awsAccessKey, awsSecretKey, region));
         }
 
-        private static class BasicKinesisProvider implements KinesisClientProvider {
+        private static final class BasicKinesisProvider implements KinesisClientProvider {
 
             private final String accessKey;
             private final String secretKey;
