--- sdks/java/io/kinesis/src/main/java/org/apache/beam/sdk/io/kinesis/source/checkpoint/generator/CheckpointGenerator.java
@@ -1,34 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.beam.sdk.io.kinesis.source.checkpoint.generator;
-
-
-import org.apache.beam.sdk.io.kinesis.client.SimplifiedKinesisClient;
-import org.apache.beam.sdk.io.kinesis.client.TransientKinesisException;
-import org.apache.beam.sdk.io.kinesis.source.checkpoint.KinesisReaderCheckpoint;
-
-import java.io.Serializable;
-
-/**
- * Used to generate checkpoint object on demand.
- * How exactly the checkpoint is generated is up to implementing class.
- */
-public interface CheckpointGenerator extends Serializable {
-    KinesisReaderCheckpoint generate(SimplifiedKinesisClient client)
-            throws TransientKinesisException;
-}
